Start the even listener before the server
I think I started the server but not in a goroutine so no events were
ever processed. I could have also used server.GoListenAndServe, but
this essentially does the same thing and feels like more control lies
the in caller.

Not sure if that is true or not, but that is where I'm going with it ;)